requirements.txt
asgiref==3.2.7
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-unixtimestampfield==0.4.0
djangorestframework==3.11.0
future==0.18.2
idna==2.9
pkg-resources==0.0.0
pythonping==1.0.9
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
textfsm==1.1.0
urllib3==1.25.8
